Restructuring as package
