+++ FabNESO/__init__.py
@@ -1,0 +1,6 @@
+"""Neptune Exploratory SOftware (NESO) plugin for FabSim3."""
+
+from .tasks import neso, neso_ensemble
+
+
+__all__ = ["neso", "neso_ensemble"]
